Initial work on port to @atomico/router
useHead will need to be updated. Something has gone mighty unstable.
