+++ package.json
@@ -79,6 +79,7 @@
   "dependencies": {
     "@a11y/skip-navigation": "0.0.10",
     "@atomico/hooks": "3.33.3",
+    "@atomico/router": "^1.1.0",
     "@emotion/css": "^11.7.1",
     "@fortawesome/free-brands-svg-icons": "6.1.1",
     "@fortawesome/free-solid-svg-icons": "6.1.1",
