+++ src/pages/Home.tsx
@@ -145,4 +145,6 @@ function component() {
   );
 }
 
-customElements.define("kepler-home", c(component));
+export const HomePage = c(component);
+
+customElements.define("kepler-home", HomePage);
